+++ src/main/kotlin/com/dsoftware/ghmanager/data/providers/DataProvider.kt
@@ -1,51 +1,48 @@
 package com.dsoftware.ghmanager.data.providers
 
+import com.dsoftware.ghmanager.api.GhApiRequestExecutor
+import com.intellij.collaboration.async.CompletableFutureUtil.submitIOTask
+import com.intellij.collaboration.util.ProgressIndicatorsProvider
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.util.EventDispatcher
 import org.jetbrains.plugins.github.api.GithubApiRequest
-import org.jetbrains.plugins.github.api.GithubApiRequestExecutor
 import org.jetbrains.plugins.github.exceptions.GithubStatusCodeException
-import org.jetbrains.plugins.github.util.LazyCancellableBackgroundProcessValue
 import java.io.IOException
 import java.util.EventListener
-import java.util.concurrent.CompletableFuture
-import kotlin.properties.ReadOnlyProperty
 
 open class DataProvider<T>(
-    private val requestExecutor: GithubApiRequestExecutor,
+    private val requestExecutor: GhApiRequestExecutor,
     private val githubApiRequest: GithubApiRequest<T>,
     private val errorValue: T?,
 ) {
     private val runChangesEventDispatcher = EventDispatcher.create(DataProviderChangeListener::class.java)
+    private val progressManager = ProgressManager.getInstance()
+    private val indicatorsProvider: ProgressIndicatorsProvider = ProgressIndicatorsProvider()
 
-    private val processValue: LazyCancellableBackgroundProcessValue<T> =
-        LazyCancellableBackgroundProcessValue.create(ProgressManager.getInstance()) {
-            try {
-                LOG.info("Executing ${githubApiRequest.url}")
-                val request = githubApiRequest
-                val response = requestExecutor.execute(it, request)
-                response
-            } catch (e: GithubStatusCodeException) {
-                LOG.warn("Error when getting $githubApiRequest.url: $e")
-                errorValue ?: throw e
-            } catch (ioe: IOException) {
-                LOG.warn("Error when getting $githubApiRequest.url: $ioe")
-                errorValue ?: throw ioe
-            }
+    private val processValue = progressManager.submitIOTask(indicatorsProvider, true) {
+        try {
+            LOG.info("Executing ${githubApiRequest.url}")
+            val request = githubApiRequest
+            val response = requestExecutor.execute(it, request)
+            response
+        } catch (e: GithubStatusCodeException) {
+            LOG.warn("Error when getting $githubApiRequest.url: status code ${e.statusCode}: ${e.message}")
+            errorValue ?: throw e
+        } catch (ioe: IOException) {
+            LOG.warn("Error when getting $githubApiRequest.url: $ioe")
+            errorValue ?: throw ioe
         }
+    }
 
-    val request by backgroundProcessValue(processValue)
-
-    private fun <T> backgroundProcessValue(backingValue: LazyCancellableBackgroundProcessValue<T>)
-        : ReadOnlyProperty<Any?, CompletableFuture<T>> =
-        ReadOnlyProperty { _, _ -> backingValue.value }
+    val request
+        get() = processValue
 
     fun url(): String = githubApiRequest.url
 
     fun reload() {
-        processValue.drop()
+        processValue.cancel(true)
         runChangesEventDispatcher.multicaster.changed()
     }