+++ src/main/kotlin/com/dsoftware/ghmanager/data/providers/LogDataProvider.kt
@@ -1,11 +1,11 @@
 package com.dsoftware.ghmanager.data.providers
 
+import com.dsoftware.ghmanager.api.GhApiRequestExecutor
 import com.dsoftware.ghmanager.api.GithubApi
 import com.dsoftware.ghmanager.api.model.Job
-import org.jetbrains.plugins.github.api.GithubApiRequestExecutor
 
 class LogDataProvider(
-    requestExecutor: GithubApiRequestExecutor,
+    requestExecutor: GhApiRequestExecutor,
     job: Job
 ) : DataProvider<String>(
     requestExecutor,